Fix deployment steps and update instructions
Summary:
This patch updates app.yaml for this project to bring it up to date with
current appengine syntax.

I also add instructions for how to deploy the two bits of this project
(one on appengine, the other on toby).

Finally I remove the module-listener.yaml file, since it's not actually
used (especially since it sent me on a goose chase trying to figure out
what appengine project the slackclient was under).

Jira ticket: https://khanacademy.atlassian.net/browse/INFRA-3420

Test Plan:
I've run the deploy steps listed in the readme and the test steps that
I've included in there, so things are working as expected!

Reviewers: csilvers, sean

Reviewed By: csilvers, sean

Differential Revision: https://phabricator.khanacademy.org/D59220