Switch to 'public' graphql endpoint
Summary:

The `/graphql` endpoint that sends traffic directly to graphql-gateway
broke due to a rollout of the render gateway service.

Previously this bot specifically sent requests to the graphql-gateway
service since we hadn't figured out how to direct traffic to go
services. But now we do that and so we can use the 'normal'
/api/internal/graphql endpoint, which will likely be more stable going
forward.

(I'll fix hte render gateway issue in a separate diff)

Test plan:
deploy and upvote some testimonials

Auditors: benkraft