Stop using no longer supported slack api call
Summary:

Laurie noticed the bot wasn't working and on investigation it looks to
be that the `channels.history` api call is now deprecated [1]. Looking
on that page, it seems like `conversations.history` is the recommended
alternative. I am not sure that these are equivalent calls, but
blindling changing these parameters seem to have worked manually testing
on toby, so going to go ahead and ship it!

[1] https://api.slack.com/changelog/2020-01-deprecating-antecedents-to-the-conversations-api

Test plan:
Deploy this code to toby, upvote some testimonials and see them echoed
to the main channel.

Auditors: benkraft, drosile