The new slack library has WebClient, not SlackClient.
I didn't realize this before because the code I was testing uses
alertlib to send to slack, not the slack library.  But when I tried
testing ./listener.py I saw the error.

This isn't a total fix because I'm getting an "invalid_auth" error,
but I was getting that with the old version of this code too so we're
not any worse off, at lesat.

Issue: none

Test plan:
env PYTHONPATH=lib ./listener.py